Added password protection
Added public key encryption on client side for the password, the password will be encrypted with a public key before sent to the server, the server will now decrypt the password from the client and from the database, and then compare if they are the same. The decryption is done with a private key hidden in the .env file.