Fixed the game page so that it works with the new db layout.
Changed from using userID to username since this reflects the changes to the db.
Updated queries to use the Username column instead of the UserID column in the db.
Fixed ambiguous column name issue.
Just commented out the settings in the game page for now since enabling them crashes the server.